requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
google-generativeai
httpx
pydantic
python-dotenv
pypdf
python-multipart
fpdf2